em: preserve double precision in carried refinement state

+++ recovar/em/dense_single_volume/relion_replay.py
@@ -657,12 +657,12 @@ def apply_iter_replay_overrides(
             _relion_offset_range,
             _relion_offset_step,
             n_classes=n_classes,
-        ).astype(np.float32)
+        ).astype(runtime_dtype)
         _state_prior_translations = _il._translation_grid_for_class_count(
             float(state.translation_range),
             float(state.translation_step),
             n_classes=n_classes,
-        ).astype(np.float32)
+        ).astype(runtime_dtype)
         _translation_grid_differs = _state_prior_translations.shape != _replay_prior_translations_np.shape
         if not _translation_grid_differs:
             _translation_grid_differs = not np.allclose(
@@ -836,9 +836,9 @@ def apply_iter_replay_overrides(
                         if not os.path.exists(_prior_star):
                             continue
                     _relion_direction_prior = (
-                        read_relion_direction_priors(_prior_star, n_classes)
+                        read_relion_direction_priors(_prior_star, n_classes, dtype=runtime_dtype)
                         if k_class_enabled
-                        else read_relion_direction_prior(_prior_star)
+                        else read_relion_direction_prior(_prior_star, dtype=runtime_dtype)
                     )
                     if k_class_enabled:
                         inferred_weights = class_weights_from_direction_prior(_relion_direction_prior, n_classes)
